# HexString: optimize performance; MacAddress: optimize putTo BVS-5367 LOXI-77