Reviewer: @ronaldchl @Sovietaced 

While debugging a performance problem related to `MacAddress.toString`, I noticed that the `HexString` conversion code we had was really old and inefficient. This code was even using anti-patterns like concatenating Strings without a StringBuilder (!).

These routines are used in a quite a lot of serialization cases, e.g., for DatapathIds.

This commit replaces the code with optimized implementations. These are necessarily relatively ugly and procedural.

JMH benchmarks show the new versions to be between 3.5 and 9.1x faster.

JMH Benchmark results:
Old versions:
```
Benchmark                                    Mode  Cnt        Score       Error  Units
HexStringBenchmark.testBytes6ToHexString    thrpt  200  5482128.668 ± 10011.236  ops/s
HexStringBenchmark.testBytes8ToHexString    thrpt  200  4097441.103 ±  9869.569  ops/s
HexStringBenchmark.testLongToHexStringPad6  thrpt  200  6015404.366 ± 16095.163  ops/s
HexStringBenchmark.testLongToHexStringPad8  thrpt  200  4499879.492 ±  7371.526  ops/s
HexStringBenchmark.testMacToString          thrpt  200  6039423.717 ± 14960.774  ops/s
HexStringBenchmark.testToBytes              thrpt  200  4320243.806 ± 12942.870  ops/s
HexStringBenchmark.testToLong               thrpt  200  5733672.143 ± 26865.832  ops/s
```

New versions:
```
Benchmark                                    Mode  Cnt         Score        Error  Units
HexStringBenchmark.testBytes6ToHexString    thrpt  200  46454822.004 ± 131143.806  ops/s
HexStringBenchmark.testBytes8ToHexString    thrpt  200  37602862.513 ±  92041.035  ops/s
HexStringBenchmark.testLongToHexStringPad6  thrpt  200  38286359.083 ±  80960.505  ops/s
HexStringBenchmark.testLongToHexStringPad8  thrpt  200  30273992.729 ±  55125.764  ops/s
HexStringBenchmark.testMacToString          thrpt  200  37012667.642 ±  93106.031  ops/s
HexStringBenchmark.testToBytes              thrpt  200  15237921.501 ±  16622.545  ops/s
HexStringBenchmark.testToLong               thrpt  200  30268266.435 ±  24626.181  ops/s
```