@andi-bigswitch 2016-05-19
@ronaldchl @Sovietaced - I've tucked on another optimization, this time for `MacAddress.putTo` (see fd6c3f5)