@ronaldchl 2016-05-31
I was wondering if my understanding of throughput (thrpt) and ops/s was inverted, until I noticed the extra digit :)