@andi-bigswitch 2016-05-31
> I was wondering if my understanding of throughput (thrpt) and ops/s was inverted, until I noticed the extra digit :)

I fell into the same trap wondering how the hell my super-duper optimized code could have ended up *slower* than the baseline... and then I noticed the digit and was considerably happier :)