@andi-bigswitch 2016-06-06
@ronaldchl @3mp4y5 @Sovietaced - this is ready for re-review/merge if you guys are happy.