@bsn-abat 2016-06-07
ABAT: START_MERGE:  Log file at http://vss1/abat/2016.06.07.1631-m.loxigen.master/abat.log