@andi-bigswitch 2016-06-07
(note: loxigen merge is fine because it was already branched off a couple of weeks ago)