@bsn-abat 2016-06-07
ABAT: ACCEPT:  Successfully merged

In case you want to see the build log, check out :
Log file at http://vss1/abat/2016.06.07.1631-m.loxigen.master/abat.log
/cc